refactor(ingest): extract WikiSource + identity helpers to wiki/source.py
